Move scabbard to its own crate
Moves the scabbard service to its own crate located at
`services/scabbard/libscabbard`.

Also includes the following changes:
- Move the scabbard-related protocol versions from `splinter::protocol`
into the scabbard crate.
- Put the service itself into a `service` submodule of the crate.
- Reorganize some imports in the new scabbard crate.
- Copy some of splinter's private hex functionality into the scabbard
crate so it's usable by scabbard.
- Move test code from the `splinter::service::test` module to the
scabbard crate, since it's no longer used in `splinter`.
- Update `splinterd`, `scabbard-cli`, and `gameroomd` to use the
`scabbard` crate.
- Rename the `scabbard-client` feature in the scabbard crate to just
`client`, and remove the `scabbard-client` feature from libsplinter.
