+++ ci/splinter-dev
@@ -71,7 +71,7 @@ RUN USER=root cargo new --bin examples/private_xo
 # Create empty Cargo projects for splinter services
 RUN USER=root cargo new --bin services/health
 RUN USER=root cargo new --bin services/scabbard/cli
-RUN USER=root cargo new --bin services/scabbard/libscabbard
+RUN USER=root cargo new --lib services/scabbard/libscabbard
 
 # Copy over splinter files
 COPY Cargo.toml /build/Cargo.toml
@@ -83,7 +83,9 @@ COPY libsplinter/protos /build/libsplinter/protos
 COPY splinterd/Cargo.toml /build/splinterd/Cargo.toml
 COPY services/health/Cargo.toml /build/services/health/Cargo.toml
 COPY services/scabbard/cli/Cargo.toml /build/services/scabbard/cli/Cargo.toml
+COPY services/scabbard/libscabbard/build.rs /build/services/scabbard/libscabbard/build.rs
 COPY services/scabbard/libscabbard/Cargo.toml /build/services/scabbard/libscabbard/Cargo.toml
+COPY services/scabbard/libscabbard/protos /build/services/scabbard/libscabbard/protos
 
 # Copy over example Cargo.toml files
 COPY examples/gameroom/cli/Cargo.toml \
@@ -129,7 +131,9 @@ RUN rm \
 RUN find . -name 'Cargo.toml' -exec \
     sh -c 'x="{}"; rm "$x" ' \;
 RUN rm /build/libsplinter/build.rs \
-    /build/libsplinter/protos/*
+    /build/libsplinter/protos/* \
+    /build/services/scabbard/libscabbard/build.rs \
+    /build/services/scabbard/libscabbard/protos/*
 
 # Log the commit hash
 COPY .git/ /tmp/.git/
