+++ examples/gameroom/daemon/Cargo.toml
@@ -49,6 +49,7 @@ percent-encoding = "2.0"
 protobuf = "2"
 sabre-sdk = "0.5"
 sawtooth-sdk = { version = "0.4", features = ["transact-compat"] }
+scabbard = { path = "../../../services/scabbard/libscabbard", features = ["events"] }
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
