+++ services/scabbard/cli/Cargo.toml
@@ -36,7 +36,7 @@ log = "0.4"
 sabre-sdk = "0.5"
 sawtooth-sdk = { version = "0.4", features = ["transact-compat"] }
 transact = { version = "0.2", features = ["contract-archive"] }
-splinter = { path = "../../../libsplinter", features = ["sawtooth-signing-compat", "scabbard-client"] }
+scabbard = { path = "../libscabbard", features = ["client"] }
 
 [dev-dependencies]
 serial_test = "0.3"
