+++ services/scabbard/libscabbard/src/client/error.rs
