+++ services/scabbard/libscabbard/src/service/error.rs
