+++ services/scabbard/libscabbard/src/service/mod.rs
@@ -16,8 +16,6 @@
 //! `transact` library for state. Scabbard uses two-phase consensus to reach agreement on
 //! transactions.
 
-#[cfg(feature = "scabbard-client")]
-pub mod client;
 mod consensus;
 mod error;
 mod factory;
@@ -34,18 +32,18 @@ use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
 use openssl::hash::{hash, MessageDigest};
-use transact::protocol::batch::BatchPair;
-use transact::protos::FromBytes;
-
-use crate::consensus::{Proposal, ProposalUpdate};
-use crate::hex::to_hex;
-use crate::protos::scabbard::{ScabbardMessage, ScabbardMessage_Type};
-use crate::signing::SignatureVerifier;
-
-use super::{
-    Service, ServiceDestroyError, ServiceError, ServiceMessageContext, ServiceNetworkRegistry,
-    ServiceStartError, ServiceStopError,
+use splinter::{
+    consensus::{Proposal, ProposalUpdate},
+    service::{
+        Service, ServiceDestroyError, ServiceError, ServiceMessageContext, ServiceNetworkRegistry,
+        ServiceStartError, ServiceStopError,
+    },
+    signing::SignatureVerifier,
 };
+use transact::{protocol::batch::BatchPair, protos::FromBytes};
+
+use super::hex::to_hex;
+use super::protos::scabbard::{ScabbardMessage, ScabbardMessage_Type};
 
 use consensus::ScabbardConsensusManager;
 use error::ScabbardError;
@@ -373,8 +371,15 @@ fn compute_db_paths(
 pub mod tests {
     use super::*;
 
-    use crate::service::tests::*;
-    use crate::signing::hash::HashVerifier;
+    use std::error::Error;
+
+    use splinter::{
+        service::{
+            ServiceConnectionError, ServiceDisconnectionError, ServiceMessageContext,
+            ServiceNetworkSender, ServiceSendError,
+        },
+        signing::hash::HashVerifier,
+    };
 
     /// Tests that a new scabbard service is properly instantiated.
     #[test]
@@ -435,5 +440,147 @@ pub mod tests {
         )
         .expect("failed to create service");
         test_connect_and_disconnect(&mut service);
+    }
+
+    #[derive(Debug)]
+    pub struct MockServiceNetworkRegistryError(pub String);
+
+    impl Error for MockServiceNetworkRegistryError {
+        fn source(&self) -> Option<&(dyn Error + 'static)> {
+            None
+        }
+    }
+
+    impl std::fmt::Display for MockServiceNetworkRegistryError {
+        fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+            write!(f, "{}", self.0)
+        }
+    }
+
+    pub struct MockServiceNetworkRegistry {
+        pub connected_ids: Arc<Mutex<HashSet<String>>>,
+        network_sender: MockServiceNetworkSender,
+    }
+
+    impl MockServiceNetworkRegistry {
+        pub fn new() -> Self {
+            MockServiceNetworkRegistry {
+                connected_ids: Arc::new(Mutex::new(HashSet::new())),
+                network_sender: MockServiceNetworkSender::new(),
+            }
+        }
+
+        pub fn network_sender(&self) -> &MockServiceNetworkSender {
+            &self.network_sender
+        }
+    }
+
+    impl ServiceNetworkRegistry for MockServiceNetworkRegistry {
+        fn connect(
+            &self,
+            service_id: &str,
+        ) -> Result<Box<dyn ServiceNetworkSender>, ServiceConnectionError> {
+            if self
+                .connected_ids
+                .lock()
+                .expect("connected_ids lock poisoned")
+                .insert(service_id.into())
+            {
+                Ok(Box::new(self.network_sender.clone()))
+            } else {
+                Err(ServiceConnectionError::RejectedError(format!(
+                    "service with id {} already connected",
+                    service_id
+                )))
+            }
+        }
+
+        fn disconnect(&self, service_id: &str) -> Result<(), ServiceDisconnectionError> {
+            if self
+                .connected_ids
+                .lock()
+                .expect("connected_ids lock poisoned")
+                .remove(service_id)
+            {
+                Ok(())
+            } else {
+                Err(ServiceDisconnectionError::RejectedError(format!(
+                    "service with id {} not connected",
+                    service_id
+                )))
+            }
+        }
+    }
+
+    #[derive(Clone, Debug)]
+    pub struct MockServiceNetworkSender {
+        pub sent: Arc<Mutex<Vec<(String, Vec<u8>)>>>,
+        pub sent_and_awaited: Arc<Mutex<Vec<(String, Vec<u8>)>>>,
+        pub replied: Arc<Mutex<Vec<(ServiceMessageContext, Vec<u8>)>>>,
+    }
+
+    impl MockServiceNetworkSender {
+        pub fn new() -> Self {
+            MockServiceNetworkSender {
+                sent: Arc::new(Mutex::new(vec![])),
+                sent_and_awaited: Arc::new(Mutex::new(vec![])),
+                replied: Arc::new(Mutex::new(vec![])),
+            }
+        }
+    }
+
+    impl ServiceNetworkSender for MockServiceNetworkSender {
+        fn send(&self, recipient: &str, message: &[u8]) -> Result<(), ServiceSendError> {
+            self.sent
+                .lock()
+                .expect("sent lock poisoned")
+                .push((recipient.to_string(), message.to_vec()));
+            Ok(())
+        }
+
+        fn send_and_await(
+            &self,
+            recipient: &str,
+            message: &[u8],
+        ) -> Result<Vec<u8>, ServiceSendError> {
+            self.sent_and_awaited
+                .lock()
+                .expect("sent_and_awaited lock poisoned")
+                .push((recipient.to_string(), message.to_vec()));
+            Ok(vec![])
+        }
+
+        fn reply(
+            &self,
+            message_origin: &ServiceMessageContext,
+            message: &[u8],
+        ) -> Result<(), ServiceSendError> {
+            self.replied
+                .lock()
+                .expect("replied lock poisoned")
+                .push((message_origin.clone(), message.to_vec()));
+            Ok(())
+        }
+
+        fn clone_box(&self) -> Box<dyn ServiceNetworkSender> {
+            Box::new(self.clone())
+        }
+    }
+
+    /// Verifies that the given service connects on start and disconnects on stop.
+    pub fn test_connect_and_disconnect(service: &mut dyn Service) {
+        let registry = MockServiceNetworkRegistry::new();
+        service.start(&registry).expect("failed to start engine");
+        assert!(registry
+            .connected_ids
+            .lock()
+            .expect("connected_ids lock poisoned")
+            .contains(service.service_id()));
+        service.stop(&registry).expect("failed to stop engine");
+        assert!(registry
+            .connected_ids
+            .lock()
+            .expect("connected_ids lock poisoned")
+            .is_empty());
     }
 }
