+++ services/scabbard/libscabbard/src/service/rest_api/mod.rs
