+++ services/scabbard/libscabbard/src/service/rest_api/resources/mod.rs
