+++ services/scabbard/libscabbard/src/service/rest_api/resources/state.rs
