+++ splinterd/Dockerfile-installed-bionic
@@ -28,7 +28,9 @@ COPY libsplinter/protos /build/libsplinter/protos
 COPY splinterd/Cargo.toml /build/splinterd/Cargo.toml
 COPY services/health/Cargo.toml /build/services/health/Cargo.toml
 COPY services/scabbard/cli/Cargo.toml /build/services/scabbard/cli/Cargo.toml
+COPY services/scabbard/libscabbard/build.rs /build/services/scabbard/libscabbard/build.rs
 COPY services/scabbard/libscabbard/Cargo.toml /build/services/scabbard/libscabbard/Cargo.toml
+COPY services/scabbard/libscabbard/protos /build/services/scabbard/libscabbard/protos
 
 # Copy over example Cargo.toml files
 COPY examples/gameroom/cli/Cargo.toml \
@@ -48,6 +50,7 @@ COPY cli /build/cli
 COPY libsplinter /build/libsplinter
 COPY services/health /build/services/health
 COPY splinterd/ /build/splinterd
+COPY services/scabbard/libscabbard /build/services/scabbard/libscabbard
 
 # Build splinterd
 WORKDIR /build/splinterd
